// src/enums/Activity.ts
/**
 * Enum representing activity types for logging user actions
 */
export enum ActivityType {
  INITIAL_PAGE = 'INITIAL_PAGE',
  LOGIN = 'LOGIN',
  LOGOUT = 'LOGOUT',
  EDIT = 'EDIT',
  CREATE = 'CREATE',
  APPROVE = 'APPROVE',
  REJECT = 'REJECT',
  CANCEL = 'CANCEL',
  DECLINE = 'DECLINE',
  ADD = 'ADD',
  SUBMIT = 'SUBMIT',
  SAVE = 'SAVE',
  DOWNLOAD = 'DOWNLOAD',
  PREVIEW = 'PREVIEW',
  VIEW = 'VIEW',
  RETURN_TO_MAKER = 'RETURN_TO_MAKER',
  RETURN_TO_TL = 'RETURN_TO_TL',
  RETURN_TO_STAFF = 'RETURN_TO_STAFF',
  RETURN_TO_ANALYST = 'RETURN_TO_ANALYST',
  RETURN_TO_SPECIALIST_DELST = 'RETURN_TO_SPECIALIST_DELST',
  ASK_FOR_INFO = 'ASK_FOR_INFO',
  ASSIGN = 'ASSIGN',
  REASSIGN = 'REASSIGN',
  NO_CHANGE = 'NO_CHANGE',
  DELETE = 'DELETE',
  COMPLETE = 'COMPLETE',
  SEND_TO_DRD = 'SEND_TO_DRD',
  AUTO_SAVE = 'AUTO_SAVE'
}

export default ActivityType;